// types/unions/schema.tags.ts
export type TagID = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12 | 13;

export type TagLabel =
  | "personal"
  | "home"
  | "family"
  | "work"
  | "urgent"
  | "high priority"
  | "low priority"
  | "health"
  | "shopping"
  | "finance"
  | "study"
  | "social"
  | "travel";

export type TagType = "context" | "priority" | "category";